explicit float -> int convertion to prevent warnings under php 8.1

+++ src/lib/gd_image.inc.php
@@ -931,7 +931,7 @@ class Image {
                     // Do nothing the text is drawn at baseline by default
                 }
             }
-            ImageTTFText ($this->img, $this->font_size, $dir, $x, $y,
+            ImageTTFText ($this->img, $this->font_size, $dir, (int)$x, (int)$y,
                           $this->current_color,$this->font_file,$txt);
 
             // Calculate and return the co-ordinates for the bounding box
@@ -1038,7 +1038,7 @@ class Image {
                 $xl -= $bbox[0]/2;
                 $yl = $y - $yadj;
                 //$xl = $xl- $xadj;
-                ImageTTFText($this->img, $this->font_size, $dir, $xl, $yl-($h-$fh)+$fh*$i,
+                ImageTTFText($this->img, $this->font_size, $dir, (int)$xl, (int)($yl-($h-$fh)+$fh*$i),
                              $this->current_color,$this->font_file,$tmp[$i]);
 
                // echo "xl=$xl,".$tmp[$i]." <br>";
